+++ src/main/java/com/github/sgoeschl/freemarker/cli/util/ObjectUtils.java
@@ -25,5 +25,4 @@ public class ObjectUtils {
     public static boolean isNotEmpty(String value) {
         return !isNullOrEmtpty(value);
     }
-
 }